+++ app/src/main/java/ru/wizand/newpasswordmanager/model/PasswordEntry.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "password_entries")
 data class PasswordEntry(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0L,  // Long для совместимости с @Insert
     val siteName: String,
     val username: String,
     val password: String
